Change check collision algorithm - up and down rects. Change encoding of source files to UTF-8.
Сменил алгоритм проверки столкновения - с верхней и нижней поверхностями.
Сменил кодировку файлов на UTF-8

+++ BallMgr.h
@@ -15,7 +15,7 @@ private:
     int dir_angle;
     int checkCollisionAngles[3];
     float dir_angle_cos, dir_angle_sin;
-    int dir_line_len;
+    float dir_line_len;
     std::vector<double> CosForAngles, SinForAngles;
     bool ball_moving;
